pyproject.toml
[project]
name = "iris"
version = "0.1.0"
description = "IRIS: Calcium-Assisted Spike Identity. Construction of optically-grounded ground-truth datasets for spike sorting benchmarks via simultaneous calcium imaging and multi-electrode array recording."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Phillip Lavrador", email = "philliplavrador@ucsb.edu" },
]
maintainers = [
    { name = "Phillip Lavrador", email = "philliplavrador@ucsb.edu" },
]
keywords = [
    "neuroscience",
    "electrophysiology",
    "calcium-imaging",
    "spike-sorting",
    "multi-electrode-array",
    "ground-truth",
    "benchmark",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "ipympl>=0.9",
    "torch>=2.0",
    "spikeinterface>=0.100",
    "h5py>=3.8",
    "tqdm>=4.60",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
daemon = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]
dev = [
    "pytest>=8",
    "pytest-cov",
    "ruff>=0.5",
    "pre-commit",
]

[project.scripts]
iris = "iris.cli:main"
iris-daemon = "iris.daemon.app:run"

[project.urls]
Homepage      = "https://github.com/philliplavrador/IRIS"
Repository    = "https://github.com/philliplavrador/IRIS"
Documentation = "https://github.com/philliplavrador/IRIS/tree/main/docs"
Issues        = "https://github.com/philliplavrador/IRIS/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/iris"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["docs/development"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q --strict-markers"
